bugfix: config parsing - was converting required floats to ints

+++ seg_lapa/config_parse/trainer_conf.py
@@ -21,12 +21,12 @@ class TrainerConf(ABC):
 @dataclass(frozen=True)
 class TrainerConfig(TrainerConf):
     gpus: int
-    overfit_batches: Union[int, float]
+    overfit_batches: Union[float, int]
     distributed_backend: Optional[str]
     precision: int
-    limit_train_batches: Union[int, float]
-    limit_val_batches: Union[int, float]
-    limit_test_batches: Union[int, float]
+    limit_train_batches: Union[float, int]
+    limit_val_batches: Union[float, int]
+    limit_test_batches: Union[float, int]
     fast_dev_run: Union[int, bool]
     max_epochs: int
     resume_from_checkpoint: Optional[str]
